fix(postgres): rewrite bootstrap process
Previously, we were keeping the database running during the bootstrap
process. However, there were a few edge cases where the database was
being booted in recovery mode during the bootstrap process, which is
not intentional. I've refactored the database to *not* boot during
the bootstrap process except when absolutely necessary, like when
we ship a base backup for the very first time.