fix(contrib): allow 5 or 6 backups
If we're checking how many backups are available in the middle of a backup, we may end up with N+1
backups because the old one has not yet been deleted. This is an acceptable case, however the
better fix would be to add checks in the integration tests to determine if a backup is occurring
before checking the number of backups available.