fix(rootfs): enable archive_mode before initial boot
It appears that the first WAL log is (occasionally) not being shipped to minio, and that's because
we did not enable `archive_mode = on` for the first database boot in `003_restore_from_backup.sh`,
which starts shipping WAL logs immediately after booting. We enable archive mode in 004. This is a
simple fix that just requires us enabling archive_mode in `003_restore_from_backup.sh` before we
boot up the server.