fix(Dockerfile): remove uses of semicolons
In bash syntax, the semicolon allows the previous command to fail and still run the next command in
the chain. In this case we do not want that to occur as the gunzip/tar/gzip commands are all
tightly coupled together, so if one fails the entire chain should fail.